SSL for DD/ImageNet-1k_SIMCLR/requirements.txt
jupyterlab==4.0.5
lightly==1.5.13
matplotlib==3.9.2
numpy==1.26.4
pillow==10.0.1
python-json-logger==2.0.7
pytorch-lightning==2.4.0
tensorboard==2.18.0
torch==2.5.0
torchaudio==2.5.0
torchmetrics==1.5.1
torchvision==0.20.0
tqdm==4.66.1

